requirements.txt
matplotlib==3.2.2
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
scikit_learn==1.1.3
seaborn==0.11.2
streamlit==1.15.1
textblob==0.15.3
tweepy==3.10.0
wordcloud==1.8.2.2